:gear: Modify : 확장자 변경 및 불필요한 파일 삭제
