+++ hooks/mutation/post.tsx
@@ -1,11 +1,11 @@
-import { apiKeys } from "@/lib/api-keys"
+"use client"
+
 import { postRequest } from "@/lib/protocol"
 import { useMutation } from "@tanstack/react-query"
 
-export const useAddReply = () => {
+export const useAddPost = () => {
   const { mutate, isPending, isError } = useMutation({
-    mutationFn: async (data: any) =>
-      await postRequest(`${apiKeys.reply}`, data),
+    mutationFn: async (data: any) => await postRequest("/posts", data),
     // onSuccess: () => navigate("/info"),
     onError: (error) => console.log(error),
   })
