+++ hooks/query/post.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { apiKeys } from "@/lib/api-keys"
 import { sortDiarys } from "@/lib/function"
 import { fetcher } from "@/lib/protocol"
@@ -5,13 +7,26 @@ import { ApiResError, ApiResSuccess } from "@/types/api-response"
 import { DiaryTypes } from "@/types/my-diarys"
 import { useQuery } from "@tanstack/react-query"
 
-export const usePostsDiarys = () => {
+export const usePostsDiarys = (type: string, productId?: number) => {
+  const searchParams = productId
+    ? new URLSearchParams([
+        ["custom", JSON.stringify({ product_id: productId })],
+      ])
+    : null
+
+  console.log("searchParams: ", searchParams?.toString())
+
   const { data, isPending, error, refetch } = useQuery<
     Record<string, DiaryTypes[]>
   >({
     queryKey: [apiKeys.posts],
     queryFn: async () => {
-      const res = await fetcher(`${apiKeys.posts}?type=seller`)
+      const res = productId
+        ? await fetcher(
+            `${apiKeys.posts}?type=${type}&${searchParams?.toString()}`
+          )
+        : await fetcher(`${apiKeys.posts}?type=${type}`)
+
       console.log("res: ", res)
       return sortDiarys(res.item)
     },
